Deprioritize current config when no current hash
This changes config merge priority so that the current config is
appended rather than prepended to the config list when there is no
current hash.

This addresses a problem where an incoming config (with hash) and
current config (without) are identical, which was previously preferring
the current config and marking the incoming config as old, even though
they are identical.  We should instead prefer the incoming one, because
it contains a clean copy of everything we want *and* has a hash, versus
our local, dirty copy that we don't yet know the hash of.