Speed up huge contacts test with fake session ids
More than half the CPU time in the multi-part config tests is being used
generating 12k proper session IDs (via Ed keygen + conversion to X) for
the test.  This replaces the session id generation with random values
which is sufficient for this test and takes almost no time at all.