Fixed a deadlock in the SnodePool
Fixed an issue where the IRouter would lose connectivity and try to use the SnodePool to rebuild paths, eventually all nodes in the SnodePool cache could end up passing the failure threshold and it would try to refresh it's cache. This refresh was being sent to the `standard_fetcher` (the IRouter) which would end up deadlocking because it would try to use the SnodePool (loop!). Now we have a `standard_fetcher_connected` check which is done before we try to use the `standard_fetcher` in order to break this loop.