Move ProProof into session_protocol.hpp
Pro proof is part of the session protocol and is required by some of the
higher-level helper functions in the crypto layer. Because of this we were
having some circular dependency issues, Debian 12 ARM build was complaining that
session_protocol.cpp was calling out to functions in
session/config/pro.h which sits in a layer above us.

This created a circular dependency and upon re-assessment the pro proof
structures really belong in the protocol layer. What does belong in the
config section is the top-level pro config object that contains the
Session Pro rotating key pair and the proof itself.