Remove Session Pro from the user config
This is not needed anymore as each device can independently manage their
own rotating keypair and get individualised proofs that they keep to
themselves for use on the network. These proofs are identified and
revoked by the generation index hash which is shared across all proofs
that are generated from a payment so it's totally fine for a client to
have their own individual proof as the generation index hash is what's
shared between all of them.

[Gen Index Hash]
|
+----Proof A {rotating pkey1}
|
+----Proof B {rotating pkey2}
|
+----...