Fix bug in set erasing causing config dirtying
The condition in set_erase_impl was not the proper negation of set
insertion, and so the clean check was only short-circuiting when the set
existed but didn't contain the value, but *not* properly
short-circuiting when the set did not exist at all.

This fixes it, along with test case.